# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /shanda/vehicles:
    get:
      operationId: VehiclesController_findAll
      parameters:
        - name: page
          required: false
          in: query
          description: 'Page number (default: 1)'
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: 'Items per page (default: 20, max: 100)'
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
      responses:
        '200':
          description: Paginated list of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleResponseDto'
      security: &ref_1
        - bearer: []
      summary: Get all vehicles for current user
      tags: &ref_2
        - vehicles
    post:
      operationId: VehiclesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleDto'
      responses:
        '201':
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponseDto'
      security: *ref_1
      summary: Create a new vehicle
      tags: *ref_2
  /shanda/vehicles/{id}:
    get:
      operationId: VehiclesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponseDto'
      security: *ref_1
      summary: Get a specific vehicle by ID
      tags: *ref_2
    put:
      operationId: VehiclesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleDto'
      responses:
        '200':
          description: Vehicle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponseDto'
      security: *ref_1
      summary: Update a vehicle
      tags: *ref_2
    delete:
      operationId: VehiclesController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_1
      summary: Delete a vehicle
      tags: *ref_2
  /shanda/vehicles/{id}/mileage:
    patch:
      operationId: VehiclesController_updateMileage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMileageDto'
      responses:
        '200':
          description: Mileage updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponseDto'
      security: *ref_1
      summary: Update vehicle mileage
      tags: *ref_2
  /shanda/vehicles/{id}/images:
    post:
      operationId: VehiclesController_uploadImages
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Images uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponseDto'
        '201':
          description: ''
      security: *ref_1
      summary: Upload images for a vehicle
      tags: *ref_2
    delete:
      operationId: VehiclesController_deleteImage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Image deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponseDto'
      security: *ref_1
      summary: Delete an image from a vehicle
      tags: *ref_2
...
info:
  title: Shanda Automotive API
  description: Marketplace connecting vehicle owners with service providers
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:4201
    description: Local Development
  - url: https://api.shanda.com
    description: Production
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    VehicleResponseDto:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        ownerId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174001
        make:
          type: string
          example: Toyota
        model:
          type: string
          example: Camry
        year:
          type: number
          example: 2022
        vin:
          type: string
          example: 1HGBH41JXMN109186
        licensePlate:
          type: string
          nullable: true
          example: ABC123
        mileage:
          type: number
          nullable: true
          example: 45000
        imageUrls:
          example:
            - https://example.com/image1.jpg
          type: array
          items:
            type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - ownerId
        - make
        - model
        - year
        - vin
        - createdAt
        - updatedAt
    CreateVehicleDto:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: number
          minimum: 1900
        vin:
          type: string
        licensePlate:
          type: string
        color:
          type: string
        mileage:
          type: number
          minimum: 0
      required:
        - make
        - model
        - year
    UpdateVehicleDto:
      type: object
      properties:
        licensePlate:
          type: string
        color:
          type: string
        mileage:
          type: number
          minimum: 0
    UpdateMileageDto:
      type: object
      properties:
        mileage:
          type: number
          minimum: 0
          description: Current mileage in kilometers
          example: 50000
      required:
        - mileage
...